feat: add types to boards.router.js
